refactor(ops): simplify AI Contract Test Maintenance System structure
- consolidate modular Python files into single test_ranker.py module
- trim runbook.md to essential info only (overview, usage, output)
- add .gitignore for Python projects and output files
